refactor(tools): extract shared archive crate cuenv-tool-archive
The GitHub and URL tool providers each carried a near-identical
zip/tar.gz/tar.xz extraction engine (~940 lines combined) with copied
dependency sets. Consolidate into crates/tools/archive with a leaf
ArchiveError type and the superset of both behaviors:

- atomic temp-dir extraction with single-root flattening and
  preferred-name binary selection (previously URL-only; the GitHub
  provider now gains atomic + normalized extraction)
- macOS .pkg payload extraction via pkgutil/cpio (previously
  GitHub-only; the URL provider now dispatches .pkg too)
- case-insensitive archive suffix sniffing via an ArchiveKind enum

Both providers now call the shared crate and map errors to
cuenv_core::Error::tool_resolution at the boundary; their duplicated
library-detection helpers are gone, and flate2/tar/xz2/zip deps drop
out of both manifests. Extraction tests consolidated into the new
crate (20 tests, including zip/raw/query-string cases).

Validation: cargo test -p cuenv-tool-archive/-tools-github/-tools-url
(72 tests green), all-target clippy clean on all three, cargo fmt,
git diff --check.
