+++ Cargo.toml
@@ -32,6 +32,7 @@ members = [
     "crates/infisical",
     "crates/aws",
     # Tool provider crates
+    "crates/tools/archive",
     "crates/tools/oci",
     "crates/tools/github",
     "crates/tools/nix",
@@ -82,6 +83,7 @@ cuenv-gcp = { path = "crates/gcp", version = "0.54.0" }
 cuenv-infisical = { path = "crates/infisical", version = "0.54.0" }
 cuenv-aws = { path = "crates/aws", version = "0.54.0" }
 # Tool provider crates
+cuenv-tool-archive = { path = "crates/tools/archive", version = "0.54.0" }
 cuenv-tools-oci = { path = "crates/tools/oci", version = "0.54.0" }
 cuenv-tools-github = { path = "crates/tools/github", version = "0.54.0" }
 cuenv-tools-nix = { path = "crates/tools/nix", version = "0.54.0" }
