+++ crates/tools/archive/src/lib.rs
@@ -1,0 +1,557 @@
+//! Shared archive extraction for cuenv tool providers.
+//!
+//! Consolidates the zip / tar.gz / tar.xz / macOS `.pkg` extraction logic
+//! previously duplicated between the GitHub and URL tool providers:
+//!
+//! - **Atomic extraction**: archives unpack into a temporary sibling
+//!   directory that atomically replaces the destination, so failed
+//!   extractions never leave partial trees behind.
+//! - **Single-root normalization**: archives whose content lives under a
+//!   single top-level directory (`tool-1.2.3/...`) are flattened so the
+//!   destination itself becomes the tool root.
+//! - **Deterministic binary selection**: [`extract_binary`] returns the
+//!   primary binary, preferring `bin/` entries whose name matches the
+//!   destination directory (or its parent for versioned cache layouts).
+//! - **macOS `.pkg` support**: payloads are expanded with `pkgutil`/`cpio`
+//!   on macOS hosts; other hosts report a descriptive error.
+
+use flate2::read::GzDecoder;
+use std::io::{Cursor, Read};
+use std::path::{Path, PathBuf};
+use tar::Archive;
+use xz2::read::XzDecoder;
+
+#[cfg(target_os = "macos")]
+mod pkg;
+
+/// Error type for archive extraction operations.
+#[derive(Debug, thiserror::Error)]
+pub enum ArchiveError {
+    /// Underlying filesystem or stream I/O failure.
+    #[error(transparent)]
+    Io(#[from] std::io::Error),
+    /// Archive parsing or content lookup failure.
+    #[error("{0}")]
+    Extraction(String),
+}
+
+impl ArchiveError {
+    /// Create an extraction error from any displayable message.
+    #[must_use]
+    pub fn extraction(message: impl Into<String>) -> Self {
+        Self::Extraction(message.into())
+    }
+}
+
+/// Result type for archive extraction operations.
+pub type Result<T> = std::result::Result<T, ArchiveError>;
+
+/// Case-insensitive suffix check that avoids allocating.
+fn has_suffix_ignore_case(name: &str, suffix: &str) -> bool {
+    name.len() >= suffix.len() && name[name.len() - suffix.len()..].eq_ignore_ascii_case(suffix)
+}
+
+/// Determine whether a filesystem path looks like a dynamic library.
+#[must_use]
+pub fn file_looks_like_library(path: &Path) -> bool {
+    let name = path
+        .file_name()
+        .and_then(|n| n.to_str())
+        .unwrap_or_default();
+    path_looks_like_library(name)
+}
+
+/// Determine whether a path string looks like a dynamic library.
+#[must_use]
+pub fn path_looks_like_library(path: &str) -> bool {
+    has_suffix_ignore_case(path, ".dylib")
+        || has_suffix_ignore_case(path, ".so")
+        || path.to_ascii_lowercase().contains(".so.")
+        || has_suffix_ignore_case(path, ".dll")
+}
+
+/// Ensure a file is executable on Unix hosts.
+///
+/// # Errors
+///
+/// Returns an error if permissions cannot be read or written.
+pub fn ensure_executable(path: &Path) -> Result<()> {
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::PermissionsExt;
+        let mut perms = std::fs::metadata(path)?.permissions();
+        perms.set_mode(0o755);
+        std::fs::set_permissions(path, perms)?;
+    }
+
+    Ok(())
+}
+
+/// Whether an extracted tree looks like a `bin`/`lib`/`include` prefix
+/// install rather than a bare binary drop.
+#[must_use]
+pub fn looks_like_prefix_install(dest: &Path) -> bool {
+    dest.join("bin").is_dir() || dest.join("lib").is_dir() || dest.join("include").is_dir()
+}
+
+/// Locate the primary binary in a prefix-style install, preferring
+/// `bin/<tool_name>`.
+///
+/// # Errors
+///
+/// Returns an error if no binary can be found.
+pub fn find_primary_binary_in_prefix(dest: &Path, tool_name: &str) -> Result<PathBuf> {
+    let preferred = dest.join("bin").join(tool_name);
+    if preferred.exists() {
+        return Ok(preferred);
+    }
+
+    find_main_binary(dest)
+}
+
+/// Extract a binary from an archive or treat the download as a raw binary.
+///
+/// `source_name` is the asset filename or download URL; any query string is
+/// ignored when sniffing the archive format. When `binary_path` is given,
+/// only that entry is extracted; otherwise the full archive is unpacked and
+/// the primary binary located.
+///
+/// # Errors
+///
+/// Returns an error when the archive cannot be parsed, the requested entry
+/// is missing, or no binary can be located after extraction.
+pub fn extract_binary(
+    data: &[u8],
+    source_name: &str,
+    binary_path: Option<&str>,
+    dest: &Path,
+) -> Result<PathBuf> {
+    let source_path = source_name.split('?').next().unwrap_or(source_name);
+
+    match ArchiveKind::detect(source_path) {
+        ArchiveKind::Zip => extract_from_zip(data, binary_path, dest),
+        ArchiveKind::TarGz => extract_from_tar_gz(data, binary_path, dest),
+        ArchiveKind::TarXz => extract_from_tar_xz(data, binary_path, dest),
+        ArchiveKind::Pkg => extract_from_pkg(data, binary_path, dest),
+        ArchiveKind::Raw => {
+            std::fs::create_dir_all(dest)?;
+            let binary_name = Path::new(source_path)
+                .file_stem()
+                .and_then(|s| s.to_str())
+                .unwrap_or("tool");
+            let binary_dest = dest.join(binary_name);
+            std::fs::write(&binary_dest, data)?;
+            ensure_executable(&binary_dest)?;
+            Ok(binary_dest)
+        }
+    }
+}
+
+/// Archive format sniffed from a source filename.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+enum ArchiveKind {
+    Zip,
+    TarGz,
+    TarXz,
+    Pkg,
+    Raw,
+}
+
+impl ArchiveKind {
+    fn detect(source_path: &str) -> Self {
+        const SUFFIXES: &[(&str, ArchiveKind)] = &[
+            (".zip", ArchiveKind::Zip),
+            (".tar.gz", ArchiveKind::TarGz),
+            (".tgz", ArchiveKind::TarGz),
+            (".tar.xz", ArchiveKind::TarXz),
+            (".txz", ArchiveKind::TarXz),
+            (".pkg", ArchiveKind::Pkg),
+        ];
+
+        SUFFIXES
+            .iter()
+            .find(|(suffix, _)| has_suffix_ignore_case(source_path, suffix))
+            .map_or(Self::Raw, |(_, kind)| *kind)
+    }
+}
+
+/// Extract from a zip archive.
+///
+/// # Errors
+///
+/// Returns an error when the archive cannot be parsed or the requested
+/// entry is missing.
+pub fn extract_from_zip(data: &[u8], binary_path: Option<&str>, dest: &Path) -> Result<PathBuf> {
+    let cursor = Cursor::new(data);
+    let mut archive = zip::ZipArchive::new(cursor)
+        .map_err(|e| ArchiveError::extraction(format!("Failed to open zip: {e}")))?;
+
+    if let Some(path) = binary_path {
+        for i in 0..archive.len() {
+            let mut file = archive
+                .by_index(i)
+                .map_err(|e| ArchiveError::extraction(format!("Failed to read zip entry: {e}")))?;
+
+            let name = file.name().to_string();
+            if name == path || name.ends_with(&format!("/{path}")) {
+                std::fs::create_dir_all(dest)?;
+                let file_name = Path::new(&name)
+                    .file_name()
+                    .and_then(|s| s.to_str())
+                    .unwrap_or(path);
+                let dest_path = dest.join(file_name);
+
+                let mut content = Vec::new();
+                file.read_to_end(&mut content)?;
+                std::fs::write(&dest_path, &content)?;
+                ensure_executable(&dest_path)?;
+
+                return Ok(dest_path);
+            }
+        }
+
+        return Err(ArchiveError::extraction(format!(
+            "Binary '{path}' not found in zip archive"
+        )));
+    }
+
+    let temp_dir = temp_extract_dir(dest);
+
+    if temp_dir.exists() {
+        std::fs::remove_dir_all(&temp_dir)?;
+    }
+    std::fs::create_dir_all(&temp_dir)?;
+
+    let extract_result = (|| -> Result<()> {
+        for i in 0..archive.len() {
+            let mut file = archive
+                .by_index(i)
+                .map_err(|e| ArchiveError::extraction(format!("Failed to read zip entry: {e}")))?;
+
+            let outpath = match file.enclosed_name() {
+                Some(path) => temp_dir.join(path),
+                None => continue,
+            };
+
+            if file.is_dir() {
+                std::fs::create_dir_all(&outpath)?;
+            } else {
+                if let Some(p) = outpath.parent() {
+                    std::fs::create_dir_all(p)?;
+                }
+                let mut content = Vec::new();
+                file.read_to_end(&mut content)?;
+                std::fs::write(&outpath, &content)?;
+
+                #[cfg(unix)]
+                if let Some(mode) = file.unix_mode() {
+                    use std::os::unix::fs::PermissionsExt;
+                    let mut perms = std::fs::metadata(&outpath)?.permissions();
+                    perms.set_mode(mode);
+                    std::fs::set_permissions(&outpath, perms)?;
+                }
+            }
+        }
+        Ok(())
+    })();
+
+    if let Err(e) = extract_result {
+        let _ = std::fs::remove_dir_all(&temp_dir);
+        return Err(e);
+    }
+
+    finalize_extracted_tree(dest, &temp_dir)?;
+    find_main_binary(dest)
+}
+
+/// Extract from a tar.gz archive.
+///
+/// # Errors
+///
+/// Returns an error when the archive cannot be parsed or the requested
+/// entry is missing.
+pub fn extract_from_tar_gz(data: &[u8], binary_path: Option<&str>, dest: &Path) -> Result<PathBuf> {
+    let decoder = GzDecoder::new(Cursor::new(data));
+    extract_from_tar(decoder, binary_path, dest, "tar.gz")
+}
+
+/// Extract from a tar.xz archive.
+///
+/// # Errors
+///
+/// Returns an error when the archive cannot be parsed or the requested
+/// entry is missing.
+pub fn extract_from_tar_xz(data: &[u8], binary_path: Option<&str>, dest: &Path) -> Result<PathBuf> {
+    let decoder = XzDecoder::new(Cursor::new(data));
+    extract_from_tar(decoder, binary_path, dest, "tar.xz")
+}
+
+/// Extract from a macOS `.pkg` archive.
+///
+/// # Errors
+///
+/// Returns an error on non-macOS hosts, or when payload expansion fails or
+/// the requested binary is missing.
+#[cfg(target_os = "macos")]
+pub fn extract_from_pkg(data: &[u8], binary_path: Option<&str>, dest: &Path) -> Result<PathBuf> {
+    pkg::extract_from_pkg(data, binary_path, dest)
+}
+
+/// Extract from a macOS `.pkg` archive (unsupported on non-macOS hosts).
+///
+/// # Errors
+///
+/// Always returns an error on non-macOS hosts.
+#[cfg(not(target_os = "macos"))]
+pub fn extract_from_pkg(_data: &[u8], _binary_path: Option<&str>, _dest: &Path) -> Result<PathBuf> {
+    Err(ArchiveError::extraction(
+        ".pkg extraction is only supported on macOS hosts".to_string(),
+    ))
+}
+
+/// Extract from a tar stream decoded by any `Read` (shared by tar.gz/tar.xz).
+fn extract_from_tar(
+    reader: impl Read,
+    binary_path: Option<&str>,
+    dest: &Path,
+    archive_kind: &str,
+) -> Result<PathBuf> {
+    let mut archive = Archive::new(reader);
+
+    std::fs::create_dir_all(dest)?;
+
+    if let Some(path) = binary_path {
+        for entry in archive
+            .entries()
+            .map_err(|e| ArchiveError::extraction(format!("Failed to read tar: {e}")))?
+        {
+            let mut entry = entry
+                .map_err(|e| ArchiveError::extraction(format!("Failed to read tar entry: {e}")))?;
+
+            let entry_path = entry
+                .path()
+                .map_err(|e| ArchiveError::extraction(format!("Invalid path in tar: {e}")))?;
+
+            let path_str = entry_path.to_string_lossy();
+            if path_str.as_ref() == path || path_str.ends_with(&format!("/{path}")) {
+                let file_name = Path::new(path)
+                    .file_name()
+                    .and_then(|s| s.to_str())
+                    .unwrap_or(path);
+                let dest_path = dest.join(file_name);
+
+                let mut content = Vec::new();
+                entry.read_to_end(&mut content)?;
+                std::fs::write(&dest_path, &content)?;
+                ensure_executable(&dest_path)?;
+
+                return Ok(dest_path);
+            }
+        }
+
+        return Err(ArchiveError::extraction(format!(
+            "Binary '{path}' not found in {archive_kind} archive"
+        )));
+    }
+
+    let temp_dir = temp_extract_dir(dest);
+    if temp_dir.exists() {
+        std::fs::remove_dir_all(&temp_dir)?;
+    }
+    std::fs::create_dir_all(&temp_dir)?;
+
+    let extract_result = archive
+        .unpack(&temp_dir)
+        .map_err(|e| ArchiveError::extraction(format!("Failed to extract tar: {e}")));
+    if let Err(err) = extract_result {
+        let _ = std::fs::remove_dir_all(&temp_dir);
+        return Err(err);
+    }
+
+    finalize_extracted_tree(dest, &temp_dir)?;
+    find_main_binary(dest)
+}
+
+fn temp_extract_dir(dest: &Path) -> PathBuf {
+    dest.with_file_name(format!(
+        ".{}.tmp",
+        dest.file_name()
+            .and_then(|s| s.to_str())
+            .unwrap_or("extract")
+    ))
+}
+
+fn single_root_dir(dir: &Path) -> Result<Option<PathBuf>> {
+    let mut entries = std::fs::read_dir(dir)?
+        .filter_map(std::result::Result::ok)
+        .collect::<Vec<_>>();
+
+    if entries.len() != 1 {
+        return Ok(None);
+    }
+
+    let only_entry = entries.swap_remove(0);
+    let entry_path = only_entry.path();
+    if entry_path.is_dir() {
+        Ok(Some(entry_path))
+    } else {
+        Ok(None)
+    }
+}
+
+fn finalize_extracted_tree(dest: &Path, temp_dir: &Path) -> Result<()> {
+    let effective_root = single_root_dir(temp_dir)?.unwrap_or_else(|| temp_dir.into());
+    let normalized_dir = temp_dir.with_file_name(format!(
+        ".{}.normalized",
+        dest.file_name()
+            .and_then(|s| s.to_str())
+            .unwrap_or("extract")
+    ));
+
+    if normalized_dir.exists() {
+        std::fs::remove_dir_all(&normalized_dir)?;
+    }
+
+    if effective_root == temp_dir {
+        std::fs::rename(temp_dir, &normalized_dir)?;
+    } else {
+        std::fs::create_dir_all(&normalized_dir)?;
+        for entry in std::fs::read_dir(&effective_root)? {
+            let entry = entry?;
+            let source_path = entry.path();
+            let target_path = normalized_dir.join(entry.file_name());
+            std::fs::rename(source_path, target_path)?;
+        }
+        std::fs::remove_dir_all(temp_dir)?;
+    }
+
+    if dest.exists() {
+        std::fs::remove_dir_all(dest)?;
+    }
+    std::fs::rename(normalized_dir, dest)?;
+    Ok(())
+}
+
+fn dir_name(path: &Path) -> Option<String> {
+    path.file_name()
+        .and_then(|name| name.to_str())
+        .map(std::borrow::ToOwned::to_owned)
+}
+
+fn looks_like_version_dir(name: &str) -> bool {
+    let normalized = name.strip_prefix('v').unwrap_or(name);
+    normalized
+        .chars()
+        .next()
+        .is_some_and(|ch| ch.is_ascii_digit())
+        && normalized
+            .chars()
+            .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '.' | '-' | '_' | '+'))
+}
+
+fn preferred_binary_names(dir: &Path) -> Vec<String> {
+    let mut names = Vec::new();
+
+    if let Some(name) = dir_name(dir) {
+        names.push(name.clone());
+
+        if looks_like_version_dir(&name)
+            && let Some(parent_name) = dir.parent().and_then(dir_name)
+            && parent_name != name
+        {
+            names.push(parent_name);
+        }
+    }
+
+    names
+}
+
+fn sorted_files(dir: &Path) -> Result<Vec<PathBuf>> {
+    let mut files = std::fs::read_dir(dir)?
+        .filter_map(std::result::Result::ok)
+        .map(|entry| entry.path())
+        .filter(|path| path.is_file())
+        .collect::<Vec<_>>();
+
+    files.sort_by(|left, right| {
+        let left_name = left
+            .file_name()
+            .and_then(|name| name.to_str())
+            .unwrap_or_default();
+        let right_name = right
+            .file_name()
+            .and_then(|name| name.to_str())
+            .unwrap_or_default();
+        left_name.cmp(right_name)
+    });
+
+    Ok(files)
+}
+
+fn executable_files(dir: &Path) -> Result<Vec<PathBuf>> {
+    let files = sorted_files(dir)?;
+
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::PermissionsExt;
+
+        Ok(files
+            .into_iter()
+            .filter(|path| {
+                std::fs::metadata(path)
+                    .map(|meta| meta.permissions().mode() & 0o111 != 0)
+                    .unwrap_or(false)
+            })
+            .collect())
+    }
+
+    #[cfg(not(unix))]
+    {
+        Ok(files)
+    }
+}
+
+fn select_preferred_binary(
+    candidates: Vec<PathBuf>,
+    preferred_names: &[String],
+) -> Option<PathBuf> {
+    for preferred_name in preferred_names {
+        if let Some(path) = candidates.iter().find(|candidate| {
+            candidate.file_name().and_then(|name| name.to_str()) == Some(preferred_name)
+        }) {
+            return Some(path.clone());
+        }
+    }
+
+    candidates.into_iter().next()
+}
+
+/// Find the main binary in an extracted directory.
+///
+/// Prefers files under `bin/` whose name matches the directory (or its
+/// parent for versioned cache layouts), then any executable at the root.
+///
+/// # Errors
+///
+/// Returns an error when no candidate binary exists.
+pub fn find_main_binary(dir: &Path) -> Result<PathBuf> {
+    let preferred_names = preferred_binary_names(dir);
+
+    let bin_dir = dir.join("bin");
+    if bin_dir.exists()
+        && let Some(path) = select_preferred_binary(sorted_files(&bin_dir)?, &preferred_names)
+    {
+        return Ok(path);
+    }
+
+    if let Some(path) = select_preferred_binary(executable_files(dir)?, &preferred_names) {
+        return Ok(path);
+    }
+
+    Err(ArchiveError::extraction(
+        "No binary found in extracted archive".to_string(),
+    ))
+}
+
+#[cfg(test)]
+mod tests;
