+++ crates/tools/github/src/tests.rs
@@ -165,20 +165,6 @@ fn test_tool_cache_dir_different_versions() {
 // ==========================================================================
 
 #[test]
-fn test_path_looks_like_library() {
-    assert!(GitHubToolProvider::path_looks_like_library(
-        "lib/libfdb_c.dylib"
-    ));
-    assert!(GitHubToolProvider::path_looks_like_library(
-        "lib/libssl.so.3"
-    ));
-    assert!(GitHubToolProvider::path_looks_like_library(
-        "bin/sqlite3.dll"
-    ));
-    assert!(!GitHubToolProvider::path_looks_like_library("bin/fdbcli"));
-}
-
-#[test]
 fn test_cache_targets_from_source_uses_lib_for_library_extract() {
     let provider = GitHubToolProvider::new();
     let temp_dir = TempDir::new().unwrap();
@@ -616,218 +602,4 @@ fn test_release_deserialization_empty_assets() {
     let json = r#"{"tag_name": "v0.1.0", "assets": []}"#;
     let release: Release = serde_json::from_str(json).unwrap();
     assert!(release.assets.is_empty());
-}
-
-// ==========================================================================
-// tar.gz / tar.xz archive extraction tests
-// ==========================================================================
-
-fn build_tar_gz(entries: &[(&str, &[u8], u32)]) -> Vec<u8> {
-    use flate2::Compression;
-    use flate2::write::GzEncoder;
-    use std::io::Write;
-    use tar::{Builder, Header};
-
-    let mut encoder = GzEncoder::new(Vec::new(), Compression::default());
-    {
-        let mut builder = Builder::new(&mut encoder);
-        for (path, contents, mode) in entries {
-            let mut header = Header::new_gnu();
-            header.set_path(path).unwrap();
-            header.set_size(contents.len() as u64);
-            header.set_mode(*mode);
-            header.set_cksum();
-            builder.append(&header, *contents).unwrap();
-        }
-        builder.finish().unwrap();
-    }
-    encoder.flush().unwrap();
-    encoder.finish().unwrap()
-}
-
-fn build_tar_xz(entries: &[(&str, &[u8], u32)]) -> Vec<u8> {
-    use std::io::Write;
-    use tar::{Builder, Header};
-
-    let mut encoder = xz2::write::XzEncoder::new(Vec::new(), 6);
-    {
-        let mut builder = Builder::new(&mut encoder);
-        for (path, contents, mode) in entries {
-            let mut header = Header::new_gnu();
-            header.set_path(path).unwrap();
-            header.set_size(contents.len() as u64);
-            header.set_mode(*mode);
-            header.set_cksum();
-            builder.append(&header, *contents).unwrap();
-        }
-        builder.finish().unwrap();
-    }
-    encoder.flush().unwrap();
-    encoder.finish().unwrap()
-}
-
-#[test]
-fn test_extract_from_tar_gz_specific_path() {
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_gz(&[
-        ("treefmt_linux_amd64/treefmt", b"#!/bin/sh\n", 0o755),
-        ("treefmt_linux_amd64/README", b"readme\n", 0o644),
-    ]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("treefmt");
-
-    let extracted = provider
-        .extract_from_tar_gz(&data, Some("treefmt"), &dest)
-        .unwrap();
-
-    assert_eq!(extracted, dest.join("treefmt"));
-    assert!(extracted.exists());
-}
-
-#[test]
-fn test_extract_from_tar_xz_specific_path() {
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_xz(&[
-        (
-            "weaver-aarch64-apple-darwin/weaver",
-            b"#!/bin/sh\necho weaver\n",
-            0o755,
-        ),
-        ("weaver-aarch64-apple-darwin/README.md", b"readme\n", 0o644),
-    ]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("weaver");
-
-    let extracted = provider
-        .extract_from_tar_xz(&data, Some("weaver-aarch64-apple-darwin/weaver"), &dest)
-        .unwrap();
-
-    assert_eq!(extracted, dest.join("weaver"));
-    assert!(extracted.exists());
-    let contents = std::fs::read(&extracted).unwrap();
-    assert_eq!(contents, b"#!/bin/sh\necho weaver\n");
-}
-
-#[test]
-fn test_extract_from_tar_xz_extracts_all() {
-    // The github extractor doesn't flatten single-root prefixes (unlike URL).
-    // Lay the archive out with bin/weaver at the top level so
-    // `find_main_binary` can locate the primary executable.
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_xz(&[
-        ("bin/weaver", b"#!/bin/sh\necho weaver\n", 0o755),
-        ("LICENSE", b"MIT\n", 0o644),
-    ]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("weaver");
-
-    let extracted = provider.extract_from_tar_xz(&data, None, &dest).unwrap();
-
-    assert_eq!(extracted, dest.join("bin").join("weaver"));
-    assert!(extracted.exists());
-    assert!(dest.join("LICENSE").exists());
-}
-
-#[test]
-fn test_extract_from_tar_gz_extracts_all() {
-    // Parity with the tar.xz version: the github extractor doesn't flatten
-    // single-root prefixes, so we lay the archive out with bin/treefmt at
-    // the top level so `find_main_binary` can locate the primary executable.
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_gz(&[
-        ("bin/treefmt", b"#!/bin/sh\necho treefmt\n", 0o755),
-        ("LICENSE", b"MIT\n", 0o644),
-    ]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("treefmt");
-
-    let extracted = provider.extract_from_tar_gz(&data, None, &dest).unwrap();
-
-    assert_eq!(extracted, dest.join("bin").join("treefmt"));
-    assert!(extracted.exists());
-    assert!(dest.join("LICENSE").exists());
-}
-
-#[test]
-fn test_extract_from_tar_matches_path_components_not_suffix() {
-    // Regression: an entry like `notweaver/x` must NOT match the
-    // requested binary path `weaver`. The match must be anchored on
-    // a `/` boundary (or full-string equality), never a raw suffix.
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_xz(&[
-        ("notweaver/x", b"decoy\n", 0o755),
-        ("subweaver", b"another-decoy\n", 0o755),
-    ]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("weaver");
-
-    let err = provider
-        .extract_from_tar_xz(&data, Some("weaver"), &dest)
-        .unwrap_err();
-    assert!(
-        err.to_string().contains("not found"),
-        "expected 'not found' error, got: {}",
-        err
-    );
-}
-
-#[test]
-fn test_extract_from_tar_xz_missing_binary_errors() {
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_xz(&[("weaver-aarch64-apple-darwin/weaver", b"#!/bin/sh\n", 0o755)]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("weaver");
-
-    let err = provider
-        .extract_from_tar_xz(&data, Some("not/in/archive"), &dest)
-        .unwrap_err();
-    assert!(err.to_string().contains("not found"));
-}
-
-#[test]
-fn test_extract_binary_detects_tar_xz_extension() {
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_xz(&[(
-        "weaver-aarch64-apple-darwin/weaver",
-        b"#!/bin/sh\necho weaver\n",
-        0o755,
-    )]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("weaver");
-
-    let extracted = provider
-        .extract_binary(
-            &data,
-            "weaver-aarch64-apple-darwin.tar.xz",
-            Some("weaver-aarch64-apple-darwin/weaver"),
-            &dest,
-        )
-        .unwrap();
-
-    assert_eq!(extracted, dest.join("weaver"));
-    assert!(extracted.exists());
-}
-
-#[test]
-fn test_extract_binary_detects_txz_extension() {
-    let provider = GitHubToolProvider::new();
-    let data = build_tar_xz(&[(
-        "weaver-aarch64-apple-darwin/weaver",
-        b"#!/bin/sh\necho weaver\n",
-        0o755,
-    )]);
-    let temp = TempDir::new().unwrap();
-    let dest = temp.path().join("weaver");
-
-    let extracted = provider
-        .extract_binary(
-            &data,
-            "weaver.txz",
-            Some("weaver-aarch64-apple-darwin/weaver"),
-            &dest,
-        )
-        .unwrap();
-
-    assert_eq!(extracted, dest.join("weaver"));
-    assert!(extracted.exists());
 }
