+++ crates/tools/url/Cargo.toml
@@ -13,18 +13,13 @@ categories.workspace = true
 [dependencies]
 async-trait = { workspace = true }
 cuenv-core = { workspace = true }
-flate2 = { workspace = true }
+cuenv-tool-archive = { workspace = true }
 reqwest = { workspace = true, features = ["rustls-no-provider"] }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 sha2 = { workspace = true }
-tar = { workspace = true }
 tokio = { workspace = true, features = ["fs", "io-util"] }
 tracing = { workspace = true }
-# xz2 prefers system liblzma via pkg-config and falls back to a bundled C build,
-# mirroring the linking posture of flate2 elsewhere in the workspace.
-xz2 = { workspace = true }
-zip = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
