+++ crates/tools/url/src/tests.rs
@@ -1,53 +1,5 @@
 use super::*;
 use cuenv_core::tools::{Arch, Os, Platform};
-use flate2::Compression;
-use flate2::write::GzEncoder;
-use std::io::Write;
-use tar::{Builder, Header};
-use tempfile::TempDir;
-
-fn build_tar_gz(entries: &[(&str, &[u8], u32)]) -> Vec<u8> {
-    let mut encoder = GzEncoder::new(Vec::new(), Compression::default());
-    {
-        let mut builder = Builder::new(&mut encoder);
-        for (path, contents, mode) in entries {
-            let mut header = Header::new_gnu();
-            header.set_path(path).unwrap();
-            header.set_size(contents.len() as u64);
-            header.set_mode(*mode);
-            header.set_cksum();
-            builder.append(&header, *contents).unwrap();
-        }
-        builder.finish().unwrap();
-    }
-    encoder.flush().unwrap();
-    encoder.finish().unwrap()
-}
-
-fn build_tar_xz(entries: &[(&str, &[u8], u32)]) -> Vec<u8> {
-    let mut encoder = xz2::write::XzEncoder::new(Vec::new(), 6);
-    {
-        let mut builder = Builder::new(&mut encoder);
-        for (path, contents, mode) in entries {
-            let mut header = Header::new_gnu();
-            header.set_path(path).unwrap();
-            header.set_size(contents.len() as u64);
-            header.set_mode(*mode);
-            header.set_cksum();
-            builder.append(&header, *contents).unwrap();
-        }
-        builder.finish().unwrap();
-    }
-    encoder.flush().unwrap();
-    encoder.finish().unwrap()
-}
-
-fn temp_dir() -> TempDir {
-    tempfile::Builder::new()
-        .prefix("cuenv_url_provider_")
-        .tempdir()
-        .unwrap()
-}
 
 #[test]
 fn test_expand_template_version() {
@@ -113,16 +65,6 @@ fn test_expand_template_all() {
 }
 
 #[test]
-fn test_path_looks_like_library() {
-    assert!(extract::path_looks_like_library("libfoo.so"));
-    assert!(extract::path_looks_like_library("libfoo.dylib"));
-    assert!(extract::path_looks_like_library("foo.dll"));
-    assert!(extract::path_looks_like_library("libfoo.so.1"));
-    assert!(!extract::path_looks_like_library("foo"));
-    assert!(!extract::path_looks_like_library("foo.tar.gz"));
-}
-
-#[test]
 fn test_provider_name() {
     let provider = UrlToolProvider::new();
     assert_eq!(provider.name(), "url");
@@ -154,122 +96,6 @@ fn test_cannot_handle_github_source() {
         extract: vec![],
     };
     assert!(!provider.can_handle(&source));
-}
-
-#[test]
-fn test_extract_from_tar_gz_flattens_node24_prefix_layout() {
-    let data = build_tar_gz(&[
-        (
-            "node-v24.14.0-linux-x64/bin/node",
-            b"#!/bin/sh\necho node24\n",
-            0o755,
-        ),
-        (
-            "node-v24.14.0-linux-x64/bin/npm",
-            b"#!/bin/sh\necho npm24\n",
-            0o755,
-        ),
-        (
-            "node-v24.14.0-linux-x64/bin/npx",
-            b"#!/bin/sh\necho npx24\n",
-            0o755,
-        ),
-        (
-            "node-v24.14.0-linux-x64/bin/corepack",
-            b"#!/bin/sh\necho corepack24\n",
-            0o755,
-        ),
-        (
-            "node-v24.14.0-linux-x64/lib/node_modules/npm/package.json",
-            br#"{"name":"npm"}"#,
-            0o644,
-        ),
-        (
-            "node-v24.14.0-linux-x64/include/node/node.h",
-            b"#define NODE_MAJOR_VERSION 24\n",
-            0o644,
-        ),
-    ]);
-    let temp = temp_dir();
-    let dest = temp.path().join("node");
-
-    let extracted = extract::extract_from_tar_gz(&data, None, &dest).unwrap();
-
-    assert_eq!(extracted, dest.join("bin").join("node"));
-    assert!(dest.join("bin").join("npm").exists());
-    assert!(dest.join("bin").join("npx").exists());
-    assert!(dest.join("bin").join("corepack").exists());
-    assert!(
-        dest.join("lib")
-            .join("node_modules")
-            .join("npm")
-            .join("package.json")
-            .exists()
-    );
-    assert!(dest.join("include").join("node").join("node.h").exists());
-}
-
-#[test]
-fn test_extract_from_tar_gz_preserves_node25_without_corepack() {
-    let data = build_tar_gz(&[
-        (
-            "node-v25.8.1-linux-x64/bin/node",
-            b"#!/bin/sh\necho node25\n",
-            0o755,
-        ),
-        (
-            "node-v25.8.1-linux-x64/bin/npm",
-            b"#!/bin/sh\necho npm25\n",
-            0o755,
-        ),
-        (
-            "node-v25.8.1-linux-x64/bin/npx",
-            b"#!/bin/sh\necho npx25\n",
-            0o755,
-        ),
-        (
-            "node-v25.8.1-linux-x64/lib/node_modules/npm/package.json",
-            br#"{"name":"npm"}"#,
-            0o644,
-        ),
-    ]);
-    let temp = temp_dir();
-    let dest = temp.path().join("node");
-
-    let extracted = extract::extract_from_tar_gz(&data, None, &dest).unwrap();
-
-    assert_eq!(extracted, dest.join("bin").join("node"));
-    assert!(dest.join("bin").join("node").exists());
-    assert!(dest.join("bin").join("npm").exists());
-    assert!(dest.join("bin").join("npx").exists());
-    assert!(!dest.join("bin").join("corepack").exists());
-}
-
-#[test]
-fn test_extract_from_tar_gz_prefers_tool_name_for_versioned_cache_dir() {
-    let data = build_tar_gz(&[
-        (
-            "node-v24.14.0-linux-x64/bin/node",
-            b"#!/bin/sh\necho node24\n",
-            0o755,
-        ),
-        (
-            "node-v24.14.0-linux-x64/bin/npm",
-            b"#!/bin/sh\necho npm24\n",
-            0o755,
-        ),
-        (
-            "node-v24.14.0-linux-x64/bin/npx",
-            b"#!/bin/sh\necho npx24\n",
-            0o755,
-        ),
-    ]);
-    let temp = temp_dir();
-    let dest = temp.path().join("node").join("24.14.0");
-
-    let extracted = extract::extract_from_tar_gz(&data, None, &dest).unwrap();
-
-    assert_eq!(extracted, dest.join("bin").join("node"));
 }
 
 #[tokio::test]
@@ -332,113 +158,6 @@ async fn test_resolve_url_with_path() {
         }
         _ => panic!("Expected URL source"),
     }
-}
-
-#[test]
-fn test_extract_from_tar_xz_specific_path() {
-    let data = build_tar_xz(&[
-        (
-            "weaver-aarch64-apple-darwin/weaver",
-            b"#!/bin/sh\necho weaver\n",
-            0o755,
-        ),
-        (
-            "weaver-aarch64-apple-darwin/README.md",
-            b"weaver readme\n",
-            0o644,
-        ),
-    ]);
-    let temp = temp_dir();
-    let dest = temp.path().join("weaver");
-
-    let extracted =
-        extract::extract_from_tar_xz(&data, Some("weaver-aarch64-apple-darwin/weaver"), &dest)
-            .unwrap();
-
-    assert_eq!(extracted, dest.join("weaver"));
-    assert!(extracted.exists());
-    let contents = std::fs::read(&extracted).unwrap();
-    assert_eq!(contents, b"#!/bin/sh\necho weaver\n");
-}
-
-#[test]
-fn test_extract_from_tar_xz_extracts_all_and_flattens_prefix() {
-    let data = build_tar_xz(&[
-        (
-            "weaver-x86_64-unknown-linux-gnu/weaver",
-            b"#!/bin/sh\necho weaver\n",
-            0o755,
-        ),
-        ("weaver-x86_64-unknown-linux-gnu/LICENSE", b"MIT\n", 0o644),
-    ]);
-    let temp = temp_dir();
-    let dest = temp.path().join("weaver");
-
-    let extracted = extract::extract_from_tar_xz(&data, None, &dest).unwrap();
-
-    // The provider promotes the single-root directory and surfaces the
-    // first executable as the primary binary.
-    assert!(extracted.exists());
-    assert!(dest.join("weaver").exists());
-    assert!(dest.join("LICENSE").exists());
-}
-
-#[test]
-fn test_extract_from_tar_xz_missing_binary_returns_error() {
-    let data = build_tar_xz(&[(
-        "weaver-aarch64-apple-darwin/weaver",
-        b"#!/bin/sh\necho weaver\n",
-        0o755,
-    )]);
-    let temp = temp_dir();
-    let dest = temp.path().join("weaver");
-
-    let err = extract::extract_from_tar_xz(&data, Some("not/in/archive"), &dest).unwrap_err();
-    assert!(err.to_string().contains("not found"));
-}
-
-#[test]
-fn test_extract_binary_detects_tar_xz_extension() {
-    let data = build_tar_xz(&[(
-        "weaver-aarch64-apple-darwin/weaver",
-        b"#!/bin/sh\necho weaver\n",
-        0o755,
-    )]);
-    let temp = temp_dir();
-    let dest = temp.path().join("weaver");
-
-    let extracted = extract::extract_binary(
-        &data,
-        "https://example.com/weaver-aarch64-apple-darwin.tar.xz",
-        Some("weaver-aarch64-apple-darwin/weaver"),
-        &dest,
-    )
-    .unwrap();
-
-    assert_eq!(extracted, dest.join("weaver"));
-    assert!(extracted.exists());
-}
-
-#[test]
-fn test_extract_binary_detects_txz_extension() {
-    let data = build_tar_xz(&[(
-        "weaver-aarch64-apple-darwin/weaver",
-        b"#!/bin/sh\necho weaver\n",
-        0o755,
-    )]);
-    let temp = temp_dir();
-    let dest = temp.path().join("weaver");
-
-    let extracted = extract::extract_binary(
-        &data,
-        "https://example.com/weaver.txz",
-        Some("weaver-aarch64-apple-darwin/weaver"),
-        &dest,
-    )
-    .unwrap();
-
-    assert_eq!(extracted, dest.join("weaver"));
-    assert!(extracted.exists());
 }
 
 #[tokio::test]
